Added basic validations to avoid duplicate entries for shipping rate APIs

+++ app/controllers/shipping-rate.controller.js
@@ -9,6 +9,9 @@ module.exports = {
         try {
 
             const newItem = await ShippingRate.createRate(req.body.rates);
+            if (!newItem.length) {
+                return res.status(400).json(ApiResponse.error("No items were added!"));
+            }
             res.status(201).json(ApiResponse.success(newItem, 'Shipping rate added!'))
 
         } catch (error) {
@@ -19,7 +22,18 @@ module.exports = {
 
     updateRate: async (req, res) => {
         try {
+            const exists = await ShippingRate.getShippingRateByID(req.params.id);
+            if (!exists) {
+                return res.status(400).json(ApiResponse.error("Shipping ID does not exists!"));
+            }
 
+            if (!exists.status) {
+                return res.status(400).json(ApiResponse.error("Cannot update a deleted entity!"));
+            }
+
+            if (exists.country != req.body.country) {
+                return res.status(400).json(ApiResponse.error("Cannot change country!"));
+            }
             const updatedItem = await ShippingRate.updateRate(req.params.id, req.body);
             res.status(200).json(ApiResponse.success(updatedItem, 'Shipping rate updated!'))
 
@@ -32,7 +46,11 @@ module.exports = {
 
     deleteRate: async (req, res) => {
         try {
-
+            const exists = await ShippingRate.getShippingRateByID(req.params.id);
+            if (!exists) {
+                return res.status(400).json(ApiResponse.error("Shipping ID does not exists!"));
+            }
+            
             const deletedItem = await ShippingRate.deleteRate(req.params.id);
             res.status(200).json(ApiResponse.success(deletedItem, 'Shipping rate deleted!'))
